refactor: split formPage
