+++ vite.config.ts
@@ -16,20 +16,7 @@ export default defineConfig({
       include: ['src/**/*'],
       exclude: ['src/components/Layout.tsx', 'src/vite-env.d.ts'],
       all: true,
-      // lines: 80,
-      // functions: 80,
-      // branches: 80,
-      // statements: 80,
       reporter: ['text'],
     },
   },
 });
-
-// coverage: {
-//   enabled: true,
-//   provider: 'c8',
-//   reporter: ['text'],
-//   all: true,
-//  include: ['src//'],
-//  exclude: ['src/.{ts,tsx}'],
-// },
